[CodeCompletion] Move implementation of CodeCompletionResultBuilder into its own file
